Added roles check for advert

+++ src/adverts/adverts.controller.ts
@@ -2,6 +2,7 @@ import {
     Body,
     Controller,
     Delete,
+    ForbiddenException,
     Get,
     HttpCode,
     HttpStatus,
@@ -21,8 +22,8 @@ import { Advert } from './entities/advert.entity';
 import { AdvertsGetDto, AdvertsGetResponseDto, UpdateAdvertDto } from './dto/advert.dto';
 import { CreateAdvertDto } from './dto/advert.dto';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
-import { JWTUserPayload } from '../auth/jwt.strategy';
-import { ResourceEnum } from '../app/app.roles';
+import { JWTPayload, JWTUserPayload } from '../auth/jwt.strategy';
+import { ResourceEnum, RolesEnum } from '../app/app.roles';
 
 @Controller('adverts')
 @ApiTags('Adverts')
@@ -51,13 +52,50 @@ export class AdvertsController {
     }
 
     @Patch(':id')
-    updateAdvert(@Param('id', ParseUUIDPipe) id: string, @Body() body: UpdateAdvertDto): Promise<Advert> {
+    @UseGuards(JwtAuthGuard, ACGuard)
+    @UseRoles({
+        resource: ResourceEnum.ADVERT,
+        action: 'update',
+        possession: 'own',
+    })
+    async updateAdvert(
+        @Param('id', ParseUUIDPipe) id: string,
+        @Body() body: UpdateAdvertDto,
+        @Req() req: Request & JWTUserPayload
+    ): Promise<Advert> {
+        if (!(await this.isAdminOrOwner(id, req.user))) {
+            throw new ForbiddenException();
+        }
         return this.advertsService.update(id, body);
     }
 
     @Delete(':id')
     @HttpCode(HttpStatus.NO_CONTENT)
-    async deleteModel(@Param('id', ParseUUIDPipe) id: string) {
-        await this.advertsService.delete(id);
+    @UseGuards(JwtAuthGuard, ACGuard)
+    @UseRoles({
+        resource: ResourceEnum.ADVERT,
+        action: 'delete',
+        possession: 'own',
+    })
+    async deleteAdvert(@Param('id', ParseUUIDPipe) id: string, @Req() req: Request & JWTUserPayload): Promise<Advert> {
+        if (!(await this.isAdminOrOwner(id, req.user))) {
+            throw new ForbiddenException();
+        }
+        return await this.advertsService.delete(id);
+    }
+
+    private async isAdminOrOwner(advertId: string, userPayload: JWTPayload): Promise<boolean> {
+        const isOwner = await this.isOwner(advertId, userPayload);
+        const isAdmin = this.isAdmin(userPayload);
+        return isOwner || isAdmin;
+    }
+
+    private async isOwner(advertId: string, userPayload: JWTPayload): Promise<boolean> {
+        const owner = await this.advertsService.getAdvertOwner(advertId);
+        return owner === userPayload.id;
+    }
+
+    private isAdmin(userPayload: JWTPayload): boolean {
+        return userPayload.roles && userPayload.roles.some(role => role === RolesEnum.ADMIN);
     }
 }
